+++ libraries/freertos_plus/standard/freertos_plus_tcp/CMakeLists.txt
@@ -50,7 +50,7 @@ afr_test_module()
 afr_module_sources(
     ${AFR_CURRENT_MODULE}
     INTERFACE
-        "${test_dir}/iot_test_freertos_tcp.c"
+        "${test_dir}/test_freertos_tcp.c"
 )
 
 afr_module_dependencies(
