+++ libraries/freertos_plus/standard/freertos_plus_tcp/test/freertos_tcp_test_access_tcp_define.h
@@ -33,7 +33,7 @@
 #ifndef _AWS_FREERTOS_TCP_TEST_ACCESS_TCP_DEFINE_H_
 #define _AWS_FREERTOS_TCP_TEST_ACCESS_TCP_DEFINE_H_
 
-#include "iot_freertos_tcp_test_access_declare.h"
+#include "freertos_tcp_test_access_declare.h"
 
 /*-----------------------------------------------------------*/
